% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-addr.R
\name{addr_row}
\alias{addr_row}
\alias{addr_row.cell_addr}
\title{Get row from cell location or reference}
\usage{
addr_row(x, ...)

\method{addr_row}{cell_addr}(x, ...)
}
\arguments{
\item{x}{a suitable representation of cell(s) or a cell area reference}

\item{...}{further arguments passed to or from other methods}
}
\value{
integer vector
}
\description{
Get row from cell location or reference
}
\section{Methods (by class)}{
\itemize{
\item \code{cell_addr}: Method for \code{\link{cell_addr}} objects
(ca <- cell_addr(1:4, 3))
addr_row(ca)
}}

