% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{as.cell_limits.character}
\alias{as.cell_limits.character}
\title{Convert a cell range into a cell_limits object}
\usage{
\method{as.cell_limits}{character}(x)
}
\arguments{
\item{x}{character vector of length one, representing a cell range}
}
\value{
a \code{cell_limits} object, which is a list with components named
  \code{rows} and \code{cols}, each of the form \code{c(min, max}
}
\description{
Convert a cell range into a cell_limits object
}
\examples{
\dontrun{
as.cell_limits("A1")
as.cell_limits("Q24")
as.cell_limits("A1:D8")
as.cell_limits("R5C11")
as.cell_limits("R2C3:R6C9")
}
}
\keyword{internal}

