% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra-ref.R
\name{as.ra_ref}
\alias{as.ra_ref}
\title{Convert to a ra_ref object}
\usage{
as.ra_ref(x, ...)
}
\arguments{
\item{x}{something}

\item{...}{potentially other stuff}
}
\value{
a \code{\link{ra_ref}} object
}
\description{
Convert something, usually a string, into an object of class
\code{\link{ra_ref}}.
}
\examples{
as.ra_ref("D$4")
as.ra_ref("R[-4]C3")
}

