% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-limits.R
\name{as.range}
\alias{as.range}
\title{Convert a cell_limits object to a cell range}
\usage{
as.range(x, fo = c("R1C1", "A1"), sheet = NULL)
}
\arguments{
\item{x}{a cell_limits object}

\item{fo}{either \code{"R1C1"} (the default) or \code{"A1"} specifying the
cell reference format; in many contexts, it can be inferred and is optional}

\item{sheet}{logical, specifying whether worksheet name should be prepended
to the range, e.g. \code{Sheet1!A1:D4}}
}
\value{
length one character vector holding a cell range
}
\description{
Convert a cell_limits object to a cell range
}
\examples{
rgCL <- cell_limits(ul = c(1, 2), lr = c(7, 6))
as.range(rgCL)
as.range(rgCL, fo = "A1")

rgCL_ws <- cell_limits(ul = c(1, 2), lr = c(7, 6), sheet = "A Sheet")
as.range(rgCL_ws)
as.range(rgCL_ws, fo = "A1")
}

