% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-addr.R
\name{cell_addr}
\alias{cell_addr}
\title{cell_addr class}
\usage{
cell_addr(row, col)
}
\arguments{
\item{row}{integer, row, must be greater than or equal to 1}

\item{col}{integer, column, must be greater than or equal to 1}
}
\value{
a \code{cell_addr} object
}
\description{
The \code{cell_addr} class is used to hold the absolute row and column
location for one or more cells. This is in contrast to the
\code{\link{ra_ref}} class, which holds a representation of a single
absolute, relative, or mixed cell reference from, e.g., a formula.
}
\section{Reference}{


Spreadsheet Implementation Technology: Basics and Extensions
Peter Sestoft
MIT Press 2014
}
\examples{
cell_addr(4, 3)
}

