% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-addr.R
\name{cell_row}
\alias{cell_row}
\alias{cell_row.cell_addr}
\title{Get row from cell location or reference}
\usage{
cell_row(x, ...)

\method{cell_row}{cell_addr}(x, ...)
}
\arguments{
\item{x}{a suitable representation of cell(s) or a cell area reference}

\item{...}{other arguments passed along to methods}
}
\value{
integer vector
}
\description{
Get row from cell location or reference
}
\section{Methods (by class)}{
\itemize{
\item \code{cell_addr}: Method for \code{\link{cell_addr}} objects
(ca <- cell_addr(1:4, 3))
cell_row(ca)
}}

