% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellranger-package.r
\docType{package}
\name{cellranger}
\alias{cellranger}
\alias{cellranger-package}
\title{cellranger}
\description{
Helper functions to work with spreadsheets and the "A1:D10" style of cell
range specification.
}

