% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{convert_cell_limits_to_range}
\alias{convert_cell_limits_to_range}
\title{Convert a cell_limits object to a cell range}
\usage{
convert_cell_limits_to_range(x, RC = FALSE)
}
\arguments{
\item{x}{a cell_limits object}

\item{RC}{logical, requesting "R1C1" positioning notation}
}
\value{
length one character vector holding a cell range, in either A1 or
  R1C1 positioning notation
}
\description{
Convert a cell_limits object to a cell range
}
\examples{
\dontrun{
rgCL <-
  structure(list(rows = c(1, 4), cols = c(1, 3)), class = "cell_limits")
convert_cell_limits_to_range(rgCL)
convert_cell_limits_to_range(rgCL, RC = TRUE)
}
}
\keyword{internal}

