% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/A1-to-from-RC.R
\name{letter_to_num}
\alias{letter_to_num}
\title{Convert column IDs from letter representation to integer}
\usage{
letter_to_num(x)
}
\arguments{
\item{x}{character vector of letter-style column IDs (case insensitive)}
}
\value{
vector of positive integers, giving associated column numbers
}
\description{
Convert column IDs from letter representation to integer
}
\examples{
letter_to_num('Z')
letter_to_num(c('AA', 'ZZ', 'ABD', 'ZZZ'))
letter_to_num(c(NA, ''))
}

