% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1-to-from-RC.R
\name{R1C1_to_A1}
\alias{R1C1_to_A1}
\title{Convert R1C1 positioning notation to A1 notation}
\usage{
R1C1_to_A1(x)
}
\arguments{
\item{x}{vector of cell positions in R1C1 notation}
}
\value{
vector of cell positions in A1 notation
}
\description{
Convert R1C1 positioning notation to A1 notation
}
\examples{
R1C1_to_A1("R1C1")
R1C1_to_A1("R10C52")
R1C1_to_A1(c("R1C1", "R10C52"))
}

