% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra-ref.R
\name{ra_ref}
\alias{ra_ref}
\title{ra_ref class}
\usage{
ra_ref(rowRef = 1L, rowAbs = TRUE, colRef = 1L, colAbs = TRUE)
}
\arguments{
\item{rowRef}{integer, row}

\item{rowAbs}{logical indicating whether \code{rowRef} is absolute or
relative}

\item{colRef}{integer, column}

\item{colAbs}{logical indicating whether \code{colRef} is absolute or
relative}
}
\value{
a \code{ra_ref} object
}
\description{
The \code{ra_ref} class is used to represent a single relative, absolute, or
mixed cell reference, presumably found in a formula. When \code{rowAbs} is
\code{TRUE}, it means that \code{rowRef} identifies a specific row in an
absolute sense. When \code{rowAbs} is \code{FALSE}, it means that
\code{rowRef} holds a positive, zero, or negative offset relative to the
address of the cell containing the formula that contains the associated cell
reference. Ditto for \code{colAbs} and \code{colRef}.
}
\section{Reference}{


Spreadsheet Implementation Technology: Basics and Extensions
Peter Sestoft
MIT Press 2014
}
\examples{
ra_ref()
ra_ref(rowRef = 3, colRef = 2)
ra_ref(rowRef = 10, rowAbs = FALSE, colRef = 3, colAbs = TRUE)
}

