% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra-ref.R
\name{to_string}
\alias{to_string}
\title{Get string representation of cell and cell area references}
\usage{
to_string(x, fo = c("A1", "R1C1"))
}
\arguments{
\item{x}{an object of class \code{\link{ra_ref}} or blah blah}

\item{fo}{either \code{"A1"} (the default) or \code{"R1C1"} specifying the
cell reference format}
}
\value{
a character vector of length one
}
\description{
Get string representation of cell and cell area references
}
\examples{
to_string(ra_ref())
to_string(ra_ref(), fo = "R1C1")
to_string(ra_ref(rowRef = 3, colRef = 2))
to_string(ra_ref(rowRef = 10, rowAbs = FALSE, colRef = 3, colAbs = TRUE),
          fo = "R1C1")
}

