% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/A1-to-from-RC.R
\name{A1_to_RC}
\alias{A1_to_RC}
\title{Convert A1 positioning notation to R1C1 notation}
\usage{
A1_to_RC(x)
}
\arguments{
\item{x}{vector of cell positions in A1 notation}
}
\value{
vector of cell positions in R1C1 notation
}
\description{
Convert "A1" style cell references to "R1C1" style. Dollar signs will be
ignored.
}
\examples{
A1_to_RC("A1")
A1_to_RC("AZ10")
A1_to_RC("AZ$10")
A1_to_RC(c("A1", "AZ10"))
A1_to_RC(c("", NA, "Q0"))
}

