% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{A1_to_RC}
\alias{A1_to_RC}
\title{Convert A1 positioning notation to R1C1 notation}
\usage{
A1_to_RC(x)
}
\arguments{
\item{x}{vector of cell positions in A1 notation}
}
\value{
vector of cell positions in R1C1 notation
}
\description{
Convert A1 positioning notation to R1C1 notation
}
\examples{
\dontrun{
A1_to_RC("A1")
A1_to_RC("AZ10")
A1_to_RC(c("A1", "AZ10"))
}
}
\keyword{internal}

