% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-addr.R
\name{as.cell_addr}
\alias{as.cell_addr}
\alias{as.cell_addr.character}
\alias{as.cell_addr.logical}
\alias{as.cell_addr.ra_ref}
\title{Convert to a cell_addr object}
\usage{
as.cell_addr(x, fo = NULL, strict = TRUE, ...)

\method{as.cell_addr}{logical}(x, ...)

\method{as.cell_addr}{ra_ref}(x, ...)

\method{as.cell_addr}{character}(x, fo = NULL, strict = TRUE, ...)
}
\arguments{
\item{x}{a cell reference}

\item{fo}{either \code{"R1C1"} (the default) or \code{"A1"} specifying the
cell reference format; in some contexts, it can be inferred}

\item{strict}{logical, indicates that only absolute references should be
converted; defaults to \code{TRUE}}

\item{...}{other arguments passed along to methods}
}
\value{
a \code{\link{cell_addr}} object
}
\description{
Convert various representations of a cell reference into an object of class
\code{\link{cell_addr}}. Recall that \code{\link{cell_addr}} objects hold
absolute row and column location, so not all \code{\link{ra_ref}} objects or
cell reference strings can be successfully converted.
}
\section{Methods (by class)}{
\itemize{
\item \code{logical}: Convert a logical (NA, really) into a cell_addr
object ... I might design this away but need it for the moment

\item \code{ra_ref}: Convert a logical (NA, really) into a cell_addr
object ... I might design this away but need it for the moment

\item \code{character}: Convert a string representation of absolute cell
references into a \code{cell_addr} object
}}
\examples{
as.cell_addr(ra_ref())
rar <- ra_ref(2, TRUE, 5, TRUE)
as.cell_addr(rar)
as.cell_addr("$D$12")
as.cell_addr("$C$4")
as.cell_addr("R4C3")
as.cell_addr(c("R4C3", "$C$4", "$D$12"))

\dontrun{
# none of these will work because they are relative references
## TO DO: first one works but gives NA, second one errors ... make consistent?
as.cell_addr("$F2")
as.cell_addr("R[-4]C3")
}
}

