% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra-ref.R
\name{as.ra_ref}
\alias{as.ra_ref}
\alias{as.ra_ref.cell_addr}
\alias{as.ra_ref.character}
\alias{as.ra_ref_v}
\alias{as.ra_ref_v.cell_addr}
\alias{as.ra_ref_v.character}
\title{Convert to a ra_ref object}
\usage{
as.ra_ref(x, ...)

as.ra_ref_v(x, ...)

\method{as.ra_ref}{character}(x, fo = NULL, warn = TRUE, strict = TRUE,
  ...)

\method{as.ra_ref_v}{character}(x, fo = NULL, warn = TRUE, strict = TRUE,
  ...)

\method{as.ra_ref}{cell_addr}(x, ...)

\method{as.ra_ref_v}{cell_addr}(x, ...)
}
\arguments{
\item{x}{one or more cell references, as a character vector or
\code{\link{cell_addr}} object}

\item{...}{further arguments passed to or from other methods}

\item{fo}{either \code{"R1C1"} (the default) or \code{"A1"} specifying the
cell reference format; in many contexts, it can be inferred and is optional}

\item{warn}{logical, \code{TRUE} (default) requests a warning if a file or
worksheet name is found in the string, since this cannot be represented in
a \code{\link{ra_ref}} object and will be dropped.}

\item{strict}{logical, \code{TRUE} (default) indicates that, for references
in "A1" format, only absolute references should be converted. If
\code{FALSE} a purely relative "A1" reference, like B4, will be treated as
purely absolute, i.e. like $B$4. Regardless of \code{strict}, a mixed "A1"
reference will lead to \code{NA}(s) in the affected position(s).}
}
\value{
a \code{\link{ra_ref}} object, in the case of \code{as.ra_ref}, or a
  list of them, in the case of \code{as.ra_ref_v}
}
\description{
Convert various representations of a cell reference into an object of class
\code{\link{ra_ref}}.
\itemize{
\item \code{as.ra_ref} is NOT vectorized and therefore requires the input to
represent exactly one cell, i.e. be of length 1.
\item \code{as.ra_ref_v} accepts input of length >= 1 and returns a list of
\code{\link{ra_ref}} objects.
}
}
\examples{
## as.ra_ref.character()
as.ra_ref("$F$2")
as.ra_ref("R[-4]C3")
as.ra_ref("B4")
as.ra_ref("B4", strict = FALSE)
as.ra_ref("B$4")

## this is actually ambiguous! is format A1 or R1C1 format?
as.ra_ref("RC2")
## format could be specified in this case
as.ra_ref("RC2", fo = "R1C1")
as.ra_ref("RC2", fo = "A1", strict = FALSE)

## as.ra_ref_v.character()
cs <- c("$A$1", "$F$14", "B$4", "D9")
\dontrun{
## won't work because as.ra_ref methods not natively vectorized
as.ra_ref(cs)
}
## use as.ra_ref_v instead
as.ra_ref_v(cs, strict = FALSE, warn = FALSE)
## as.ra_ref.cell_addr
ca <- cell_addr(2, 5)
as.ra_ref(ca)
## as.ra_ref_v.cell_addr()

ca <- cell_addr(1:3, 1)
\dontrun{
## won't work because as.ra_ref methods not natively vectorized
as.ra_ref(ca)
}
## use as.ra_ref_v instead
as.ra_ref_v(ca)
}

