% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-addr.R, R/ra-ref.R
\name{as.ra_ref.cell_addr}
\alias{as.ra_ref}
\alias{as.ra_ref.cell_addr}
\alias{as.ra_ref.character}
\title{Convert to a ra_ref object}
\usage{
\method{as.ra_ref}{cell_addr}(x, ...)

as.ra_ref(x, ...)

\method{as.ra_ref}{character}(x, fo = NULL, warn = TRUE, strict = TRUE,
  ...)
}
\arguments{
\item{x}{a cell reference}

\item{...}{further arguments passed to or from other methods}

\item{fo}{either \code{"R1C1"} (the default) or \code{"A1"} specifying the
cell reference format; in many contexts, it can be inferred and is optional}

\item{warn}{logical, \code{TRUE} (default) requests a warning if a file or
worksheet name is found in the string, since this cannot be represented in
a \code{\link{ra_ref}} object and will be dropped.}

\item{strict}{logical, \code{TRUE} (default) indicates that, for references
in "A1" format, only absolute references should be converted. If
\code{FALSE} a purely relative "A1" reference, like B4, will be treated as
purely absolute, i.e. like $B$4. Regardless of \code{strict}, a mixed
"A1" reference will lead to \code{NA}(s) in the affected position(s).}
}
\value{
a \code{\link{ra_ref}} object
}
\description{
Convert various representations of a cell reference into an object of class
\code{\link{ra_ref}}. Note this function is NOT vectorized, but see the
examples.
}
\section{Methods (by class)}{
\itemize{
\item \code{cell_addr}: Convert a \code{cell_addr} into a \code{\link{ra_ref}}

\item \code{character}: Convert a string representation of a cell reference
into an object of class \code{\link{ra_ref}}
}}
\examples{
ca <- cell_addr(2, 5)
as.ra_ref(ca)

ca <- cell_addr(1:3, 1)
\dontrun{
## won't work because as.ra_ref methods not natively vectorized
as.ra_ref(ca)
}
## but it's easy enough to do with Vectorize
f <- Vectorize(as.ra_ref, USE.NAMES = FALSE, SIMPLIFY = FALSE)
f(ca)
as.ra_ref("$F$2")
as.ra_ref("R[-4]C3")
as.ra_ref("B4")
as.ra_ref("B4", strict = FALSE)
as.ra_ref("B$4")

## this is actually ambiguous! is format A1 or R1C1 format?
as.ra_ref("RC2")
## format could be specified in this case
as.ra_ref("RC2", fo = "R1C1")
as.ra_ref("RC2", fo = "A1", strict = FALSE)

cs <- c("$A$1", "$F$14")
\dontrun{
## won't work because as.ra_ref methods not natively vectorized
as.ra_ref(cs)
}
## but it's easy enough to do with Vectorize
f <- Vectorize(as.ra_ref, USE.NAMES = FALSE, SIMPLIFY = FALSE)
f(cs)
}

