% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-limits.R
\name{as.range}
\alias{as.range}
\title{Convert a cell_limits object to a cell range}
\usage{
as.range(x, RC = FALSE, wsn = FALSE)
}
\arguments{
\item{x}{a cell_limits object}

\item{RC}{logical, requesting "R1C1" positioning notation}

\item{wsn}{logical, specifying whether worksheet name should be prepended to
the range, e.g. \code{Sheet1!A1:D4}}
}
\value{
length one character vector holding a cell range, in either A1 or
  R1C1 positioning notation
}
\description{
Convert a cell_limits object to a cell range
}
\examples{
rgCL <- cell_limits(ul = c(1, 2), lr = c(7, 6))
as.range(rgCL)
as.range(rgCL, RC = TRUE)

rgCL_ws <- cell_limits(ul = c(1, 2), lr = c(7, 6), wsn = "A Sheet")
as.range(rgCL_ws)
as.range(rgCL_ws, RC = TRUE, wsn = TRUE)

}

