% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cell-limits.R
\name{as.range}
\alias{as.range}
\title{Convert a cell_limits object to a cell range}
\usage{
as.range(x, RC = FALSE)
}
\arguments{
\item{x}{a cell_limits object}

\item{RC}{logical, requesting "R1C1" positioning notation}
}
\value{
length one character vector holding a cell range, in either A1 or
  R1C1 positioning notation
}
\description{
Convert a cell_limits object to a cell range
}
\examples{
rgCL <- cell_limits(ul = c(1, 2), lr = c(7, 6))
as.range(rgCL)
as.range(rgCL, RC = TRUE)
}

