% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{letter_to_num}
\alias{letter_to_num}
\title{Convert column IDs from letter representation to integer}
\usage{
letter_to_num(x)
}
\arguments{
\item{x}{character vector of letter-style column IDs (case insensitive)}
}
\value{
vector of positive integers identifying columns numerically
}
\description{
Convert column IDs from letter representation to integer
}
\examples{
\dontrun{
letter_to_num('Z')
letter_to_num(c('AA', 'ZZ', 'ABD', 'ZZZ'))
}
}
\keyword{internal}

