% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{num_to_letter}
\alias{num_to_letter}
\title{Convert column IDs from numeric to letter representation}
\usage{
num_to_letter(y)
}
\arguments{
\item{x}{vector of numeric column IDs}
}
\value{
character vector of letter-style column IDs
}
\description{
Convert column IDs from numeric to letter representation
}
\examples{
\dontrun{
num_to_letter(28)
num_to_letter(900)
num_to_letter(18278)
num_to_letter(c(25, 52, 900, 18278))
}
}
\keyword{internal}

