% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cell-specification.R
\name{RC_to_A1}
\alias{RC_to_A1}
\title{Convert R1C1 positioning notation to A1 notation}
\usage{
RC_to_A1(x)
}
\arguments{
\item{x}{vector of cell positions in R1C1 notation}
}
\value{
vector of cell positions in A1 notation
}
\description{
Convert R1C1 positioning notation to A1 notation
}
\examples{
\dontrun{
A1_to_RC("A1")
A1_to_RC("AZ10")
A1_to_RC(c("A1", "AZ10"))
}
}
\keyword{internal}

