% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-addr.R, R/ra-ref.R
\name{to_string.cell_addr}
\alias{to_string}
\alias{to_string.cell_addr}
\alias{to_string.ra_ref}
\title{Get string representation of cell references}
\usage{
\method{to_string}{cell_addr}(x, fo = c("R1C1", "A1"))

to_string(x, fo = c("R1C1", "A1"))

\method{to_string}{ra_ref}(x, fo = c("R1C1", "A1"))
}
\arguments{
\item{x}{a suitable representation of a cell or cell area reference}

\item{fo}{either \code{"R1C1"} (the default) or \code{"A1"} specifying the
cell reference format}
}
\value{
a character vector of length one
}
\description{
If either the row or column reference is relative, note that, in general,
it's impossible to convert to an "A1" formatted string. We would have to know
"relative to what?".
}
\section{Methods (by class)}{
\itemize{
\item \code{cell_addr}: Convert a \code{\link{cell_addr}} object to a cell
reference string

\item \code{ra_ref}: Convert a \code{\link{ra_ref}} object to a cell
reference string
}}
\examples{
to_string(ra_ref())
to_string(ra_ref(), fo = "A1")
to_string(ra_ref(rowRef = 3, colRef = 2))
(rel_ref <- ra_ref(rowRef = 10, rowAbs = FALSE, colRef = 3))
to_string(rel_ref)
\dontrun{
## this won't work because row reference is relative and format is A1
to_string(rel_ref, fo = "A1")
}
}

