% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-string.R
\name{to_string}
\alias{to_string}
\alias{to_string.cell_addr}
\alias{to_string.ra_ref}
\title{Get string representation of cell references}
\usage{
to_string(x, fo = c("R1C1", "A1"), ...)

\method{to_string}{ra_ref}(x, fo = c("R1C1", "A1"), ...)

\method{to_string}{cell_addr}(x, fo = c("R1C1", "A1"), ...)
}
\arguments{
\item{x}{a suitable representation of a cell or cell area reference}

\item{fo}{either \code{"R1C1"} (the default) or \code{"A1"} specifying the
cell reference format; in many contexts, it can be inferred and is optional}

\item{...}{further arguments passed to or from other methods}
}
\value{
a character vector of length one
}
\description{
If either the row or column reference is relative, note that, in general,
it's impossible to convert to an "A1" formatted string. We would have to know
"relative to what?".
}
\section{Methods (by class)}{
\itemize{
\item \code{ra_ref}: Convert a \code{\link{ra_ref}} object to a cell
reference string

\item \code{cell_addr}: Convert a \code{\link{cell_addr}} object to a cell
reference string
}}
\examples{
## ra_ref --> string
to_string(ra_ref())
to_string(ra_ref(), fo = "A1")
to_string(ra_ref(rowRef = 3, colRef = 2))
(mixed_ref <- ra_ref(rowRef = 10, rowAbs = FALSE, colRef = 3))
to_string(mixed_ref)
\dontrun{
## this will raise warning and generate NA, because row reference is
## relative and format is A1
to_string(mixed_ref, fo = "A1")
}
## cell_addr --> string
(ca <- cell_addr(3, 8))
to_string(ca)
to_string(ca, fo = "A1")

(ca <- cell_addr(1:4, 3))
to_string(ca)
to_string(ca, fo = "A1")
}

